Rename utils files, set `skipLibCheck` to `false`
