+++ package.json
@@ -58,7 +58,7 @@
     "compile": "tsc -p ./",
     "lint": "eslint src --ext ts",
     "watch": "tsc -watch -p ./",
-    "pretest": "yarn run compile && yarn run lint",
+    "pretest": "yarn compile && yarn lint",
     "test": "node ./out/test/runTest.js",
     "test:unit": "yarn jest",
     "test:watch": "yarn jest --watch",
