+++ src/commands/extractTranslationString/utils/string.ts
